app/vlinsert/loki: move entities to places where those are used
